Add files via upload
I wanted to track all the various rolls made in my friends' D&D game. I track the flat D20 roll, the type of roll it is, which character, etc. I then am able to use Pivot Tables to find the average D20 roll per character, the average types of rolls per character, and more. This gives us insight into how we play our characters and which skills we prioritize in the game.